converted modules test
