+++ examples/editor/app.js
@@ -3,7 +3,7 @@ import document from 'global/document';
 import * as React from 'react';
 import ReactDOM from 'react-dom';
 
-import { Example } from './example.js';
+import { Example } from './example';
 
 const root = document.createElement('div');
