+++ jest.config.js
@@ -1,6 +1,7 @@
+require("core-js/stable");
 module.exports = {
   verbose: true,
   testURL: 'http://localhost/',
-  collectCoverageFrom: ['modules/*/src/**/*.js', '!**/node_modules/**'],
+  collectCoverageFrom: ['modules/*/src/**/*.{ts,tsx}', '!**/node_modules/**'],
   testPathIgnorePatterns: ['/node_modules/', '/website/']
 };
