+++ modules/edit-modes/test/lib/transform-mode.test.ts
@@ -1,7 +1,6 @@
-// @flow
 
 import { TransformMode } from '../../src/lib/transform-mode';
-import { createFeatureCollectionProps, createPointerMoveEvent } from '../test-utils.js';
+import { createFeatureCollectionProps, createPointerMoveEvent } from '../test-utils';
 
 let transformMode: TransformMode;
 
@@ -9,7 +8,7 @@ let warnBefore;
 beforeEach(() => {
   warnBefore = console.warn; // eslint-disable-line
   // $FlowFixMe
-  console.warn = function () {}; // eslint-disable-line
+  console.warn = function () { }; // eslint-disable-line
   transformMode = new TransformMode();
 });
