+++ modules/layers/test/lib/layers/elevated-edit-handle-layer.test.ts
@@ -1,7 +1,6 @@
-// @flow
 /* eslint-env jest */
 
-import ElevatedEditHandleLayer from '../../../src/layers/elevated-edit-handle-layer.js';
+import ElevatedEditHandleLayer from '../../../src/layers/elevated-edit-handle-layer';
 
 describe('ElevatedEditHandleLayer tests', () => {
   test('renderLayers()', () => {
