+++ modules/overlays/test/html-tooltip-overlay.test.tsx
@@ -1,4 +1,3 @@
-// @flow-ignore
 import React from 'react';
 import Renderer from 'react-test-renderer';
