+++ package.json
@@ -56,7 +56,6 @@
     "@deck.gl/layers": "^8.0.6",
     "@deck.gl/mesh-layers": "^8.0.6",
     "@deck.gl/react": "^8.0.6",
-    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "10.1.0",
     "babel-jest": "25.2.6",
     "babel-loader": "8.1.0",
@@ -83,12 +82,14 @@
     "raw-loader": "^0.5.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-test-renderer": "^16.4.1",
+    "react-test-renderer": "^16.13.1",
+    "@types/react-test-renderer": "^16.9.2",
     "@types/react": "16.9.32",
     "@types/react-dom": "16.9.6",
     "@types/geojson": "^7946.0.7",
     "regenerator-runtime": "^0.12.1",
     "reify": "0.4.4",
+    "core-js": "3.6.4",
     "source-map-loader": "^0.2.4",
     "tap-browser-color": "^0.1.2",
     "tape": "4.8",
